Continue refactoring to reduce the need to include Eigen and OpenFST headers compilation units.
