+++ .clang-tidy
@@ -1,5 +1,5 @@
 ---
-Checks:          'clang-diagnostic-*,clang-analyzer-*,cert-*,cppcoreguidelines-*,google-*,llvm-*,misc-*,modernize-*,performance-*,portability-*,readility-*,-cppcoreguidelines-pro-bounds-pointer-arithmetic,-llvm-header-guard,-llvm-include-order,-clang-diagnostic-missing-braces,-misc-misplaced-widening-cast,-cppcoreguidelines-avoid-magic-numbers,-modernize-use-trailing-return-type,-cppcoreguidelines-avoid-c-arrays,-modernize-avoid-c-arrays,-cppcoreguidelines-pro-bounds-constant-array-index'
+Checks:          'clang-diagnostic-*,clang-analyzer-*,cert-*,cppcoreguidelines-*,google-*,llvm-*,misc-*,modernize-*,performance-*,portability-*,readility-*,-cppcoreguidelines-pro-bounds-pointer-arithmetic,-llvm-header-guard,-llvm-include-order,-clang-diagnostic-missing-braces,-misc-misplaced-widening-cast,-cppcoreguidelines-avoid-magic-numbers,-modernize-use-trailing-return-type,-cppcoreguidelines-avoid-c-arrays,-modernize-avoid-c-arrays,-cppcoreguidelines-pro-bounds-constant-array-index,-misc-const-correctness,-misc-confusable-identifiers,-performance-unnecessary-value-param'
 WarningsAsErrors: ''
 HeaderFilterRegex: 'src/include/coati'
 AnalyzeTemporaryDtors: false
