+++ src/lib/fasta.cc
@@ -24,6 +24,7 @@
 
 #include <coati/fasta.hpp>
 #include <coati/mutation_fst.hpp>
+#include <coati/structs.hpp>
 
 namespace coati {
