+++ src/lib/mutation_fst.cc
@@ -26,6 +26,7 @@
 #include <coati/dna_syms.hpp>
 #include <coati/mutation_coati.hpp>
 #include <coati/mutation_fst.hpp>
+#include <coati/utils.hpp>
 
 namespace coati {
